+++ app/src/main/java/org/linphone/receivers/NotificationBroadcastReceiver.java
@@ -31,36 +31,52 @@ import org.linphone.LinphoneManager;
 import org.linphone.LinphoneService;
 import org.linphone.compatibility.Compatibility;
 import org.linphone.core.Address;
+import org.linphone.core.Call;
 import org.linphone.core.ChatMessage;
 import org.linphone.core.ChatRoom;
 import org.linphone.core.Core;
 import org.linphone.core.ProxyConfig;
+import org.linphone.core.Reason;
 
 public class NotificationBroadcastReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
-        String reply = getMessageText(intent).toString();
-        if (reply != null) {
+        if (intent.getAction() == Compatibility.INTENT_REPLY_NOTIF_ACTION) {
+            String reply = getMessageText(intent).toString();
+            if (reply == null) return;
             Notification replied = Compatibility.createRepliedNotification(context, reply);
-            if (replied != null) {
-                int notifId = intent.getIntExtra(Compatibility.INTENT_NOTIF_ID, 0);
-                String remoteSipAddr = LinphoneService.instance().getSipUriForNotificationId(notifId);
+            if (replied == null) return;
+            int notifId = intent.getIntExtra(Compatibility.INTENT_NOTIF_ID, 0);
+            String remoteSipAddr = LinphoneService.instance().getSipUriForNotificationId(notifId);
 
-                Core core = LinphoneManager.getLc();
-                if (core == null) return;
-                ProxyConfig proxyConfig = core.getDefaultProxyConfig();
-                if (proxyConfig == null) return;
-                Address localAddr = proxyConfig.getIdentityAddress();
-                Address remoteAddr = core.interpretUrl(remoteSipAddr);
-                if (localAddr == null || remoteAddr == null) return;
-                ChatRoom room = core.findChatRoom(remoteAddr, localAddr);
-                if (room == null) return;
+            Core core = LinphoneManager.getLc();
+            if (core == null) return;
+            ProxyConfig proxyConfig = core.getDefaultProxyConfig();
+            if (proxyConfig == null) return;
+            Address localAddr = proxyConfig.getIdentityAddress();
+            Address remoteAddr = core.interpretUrl(remoteSipAddr);
+            if (localAddr == null || remoteAddr == null) return;
+            ChatRoom room = core.findChatRoom(remoteAddr, localAddr);
+            if (room == null) return;
 
-                room.markAsRead();
-                ChatMessage msg = room.createMessage(reply);
-                msg.send();
+            room.markAsRead();
+            ChatMessage msg = room.createMessage(reply);
+            msg.send();
 
-                LinphoneService.instance().sendNotification(replied, notifId);
+            LinphoneService.instance().sendNotification(replied, notifId);
+        } else if (intent.getAction() == Compatibility.INTENT_ANSWER_CALL_NOTIF_ACTION || intent.getAction() == Compatibility.INTENT_HANGUP_CALL_NOTIF_ACTION) {
+            int callId = intent.getIntExtra(Compatibility.INTENT_CALL_ID, 0);
+            String remoteAddr = LinphoneService.instance().getSipUriForCallNotificationId(callId);
+
+            Core core = LinphoneManager.getLc();
+            if (core == null) return;
+            Call call = core.findCallFromUri(remoteAddr);
+            if (call == null) return;
+
+            if (intent.getAction() == Compatibility.INTENT_ANSWER_CALL_NOTIF_ACTION) {
+                call.accept();
+            } else {
+                call.decline(Reason.None);
             }
         }
     }